+++ src/parser.rs
@@ -120,7 +120,7 @@ impl GherkinEnv {
         let line = line_offsets
             .iter()
             .position(|x| x > &offset)
-            .unwrap_or_else(|| line_offsets.len());
+            .unwrap_or(line_offsets.len());
 
         let col = offset - line_offsets[line - 1] + 1;
 
@@ -153,7 +153,7 @@ impl Default for GherkinEnv {
 peg::parser! { pub(crate) grammar gherkin_parser(env: &GherkinEnv) for str {
 
 rule _() = quiet!{[' ' | '\t']*}
-rule __() = quiet!{[' ' | '\t']+}
+rule __() = quiet!{([' ' | '\t'] / nl())*}
 
 rule nl0() = quiet!{"\r"? "\n"}
 rule nl() = quiet!{nl0() p:position!() comment()* {
@@ -164,7 +164,7 @@ rule nl_eof() = quiet!{(nl() / [' ' | '\t'])+ / eof()}
 rule comment() = quiet!{[' ' | '\t']* "#" $((!nl0()[_])*) nl_eof()}
 rule not_nl() -> &'input str = n:$((!nl0()[_])+) { n }
 
-rule keyword1(list: &[&'static str]) -> &'static str
+rule keyword1(list: &[&str]) -> &'input str
     = input:$([_]*<
         {list.iter().map(|x| x.chars().count()).min().unwrap()},
         {list.iter().map(|x| x.chars().count()).max().unwrap()}
@@ -182,7 +182,7 @@ rule keyword1(list: &[&'static str]) -> &'static str
         }
     }
 
-rule keyword0(list: &[&'static str]) -> usize
+rule keyword0(list: &[&str]) -> usize
     = keyword1(list)? {?
         match env.last_keyword().as_ref() {
             Some(v) => Ok(v.chars().count()),
@@ -190,14 +190,14 @@ rule keyword0(list: &[&'static str]) -> usize
         }
     }
 
-pub(crate) rule keyword(list: &[&'static str]) -> &'static str
+pub(crate) rule keyword<'a>(list: &[&'a str]) -> &'a str
     = comment()* len:keyword0(list) [_]*<{len}> {
         let kw = env.take_keyword();
         list.iter().find(|x| **x == &*kw).unwrap()
     }
 
 rule language_directive() -> ()
-    = "#" _ "language:" _ l:$(not_nl()+) _ nl() {?
+    = __ "#" _ "language" _ ":" _ l:$(not_nl()+) _ nl() {?
         env.set_language(l)
     }
 
@@ -245,53 +245,53 @@ pub(crate) rule table() -> Table
     }
 
 pub(crate) rule step() -> Step
-    = comment()* pa:position!() k:keyword((env.keywords().given)) __ n:not_nl() pb:position!() _ nl_eof() _
+    = comment()* pa:position!() k:keyword((env.keywords().given)) _ n:not_nl() pb:position!() _ nl_eof() _
       d:docstring()? t:table()?
     {
         env.set_last_step(StepType::Given);
         Step::builder().ty(StepType::Given)
             .keyword(k.to_string())
-            .value(n.to_string())
+            .value(n.trim_end().to_string())
             .table(t)
             .docstring(d)
             .span(Span { start: pa, end: pb })
             .position(env.position(pa))
             .build()
     }
-    / pa:position!() k:keyword((env.keywords().when)) __ n:not_nl() pb:position!() _ nl_eof() _
+    / pa:position!() k:keyword((env.keywords().when)) _ n:not_nl() pb:position!() _ nl_eof() _
       d:docstring()? t:table()?
     {
         env.set_last_step(StepType::When);
         Step::builder().ty(StepType::When)
             .keyword(k.to_string())
-            .value(n.to_string())
+            .value(n.trim_end().to_string())
             .table(t)
             .docstring(d)
             .span(Span { start: pa, end: pb })
             .position(env.position(pa))
             .build()
     }
-    / pa:position!() k:keyword((env.keywords().then)) __ n:not_nl() pb:position!() _ nl_eof() _
+    / pa:position!() k:keyword((env.keywords().then)) _ n:not_nl() pb:position!() _ nl_eof() _
       d:docstring()? t:table()?
     {
         env.set_last_step(StepType::Then);
         Step::builder().ty(StepType::Then)
             .keyword(k.to_string())
-            .value(n.to_string())
+            .value(n.trim_end().to_string())
             .table(t)
             .docstring(d)
             .span(Span { start: pa, end: pb })
             .position(env.position(pa))
             .build()
     }
-    / pa:position!() k:keyword((env.keywords().and)) __ n:not_nl() pb:position!() _ nl_eof() _
+    / pa:position!() k:keyword((env.keywords().and)) _ n:not_nl() pb:position!() _ nl_eof() _
       d:docstring()? t:table()?
     {?
         match env.last_step() {
             Some(v) => {
                 Ok(Step::builder().ty(v)
                     .keyword(k.to_string())
-                    .value(n.to_string())
+                    .value(n.trim_end().to_string())
                     .table(t)
                     .docstring(d)
                     .span(Span { start: pa, end: pb })
@@ -303,14 +303,14 @@ pub(crate) rule step() -> Step
             }
         }
     }
-    / pa:position!() k:keyword((env.keywords().but)) __ n:not_nl() pb:position!() _ nl_eof() _
+    / pa:position!() k:keyword((env.keywords().but)) _ n:not_nl() pb:position!() _ nl_eof() _
       d:docstring()? t:table()?
     {?
         match env.last_step() {
             Some(v) => {
                 Ok(Step::builder().ty(v)
                     .keyword(k.to_string())
-                    .value(n.to_string())
+                    .value(n.trim_end().to_string())
                     .table(t)
                     .docstring(d)
                     .span(Span { start: pa, end: pb })
@@ -332,12 +332,14 @@ pub(crate) rule steps() -> Vec<Step>
 rule background() -> Background
     = comment()* _ pa:position!()
       k:keyword((env.keywords().background)) ":" _ n:not_nl()? nl_eof()
+      d:description((&env.keywords().excluded_background()))?
       s:steps()?
       pb:position!()
     {
         Background::builder()
             .keyword(k.into())
-            .name(n.map(str::to_string))
+            .name(n.unwrap_or_default().trim_end().to_string())
+            .description(d.flatten())
             .steps(s.unwrap_or_default())
             .span(Span { start: pa, end: pb })
             .position(env.position(pa))
@@ -349,16 +351,16 @@ rule any_directive() -> &'static str
         k
     }
 
-rule description_line() -> &'input str
+rule description_line(excluded: &[&str]) -> &'input str
     = _
-      !"@" !keyword((&*env.keywords().exclude_in_description()))
+      !"@" !keyword((excluded))
       _ n:not_nl() nl_eof()
     {
         n
     }
 
-rule description() -> Option<String>
-    = d:(description_line() ** _) {
+rule description(excluded: &[&str]) -> Option<String>
+    = d:(description_line(excluded) ** _) __ {
         let d = d.join("\n");
         if d.trim() == "" {
             None
@@ -374,12 +376,14 @@ rule examples() -> Examples
       _
       pa:position!()
       k:keyword((env.keywords().examples)) ":" _ n:not_nl()? nl_eof()
-      tb:table()
+      d:description((&env.keywords().excluded_examples()))?
+      tb:table()?
       pb:position!()
     {
         Examples::builder()
             .keyword(k.into())
-            .name(n.map(str::to_owned))
+            .name(n.map(|n| n.trim_end().to_string()))
+            .description(d.flatten())
             .tags(t)
             .table(tb)
             .span(Span { start: pa, end: pb })
@@ -393,14 +397,16 @@ rule scenario() -> Scenario
       t:tags()
       _
       pa:position!()
-      k:keyword((env.keywords().scenario)) ":" _ n:not_nl() _ nl_eof()
+      k:keyword((env.keywords().scenario)) ":" _ n:not_nl()? _ nl_eof()
+      d:description((&env.keywords().excluded_scenario()))?
       s:steps()?
       e:examples()*
       pb:position!()
     {
         Scenario::builder()
             .keyword(k.into())
-            .name(n.to_string())
+            .name(n.unwrap_or_default().trim_end().to_string())
+            .description(d.flatten())
             .tags(t)
             .steps(s.unwrap_or_default())
             .examples(e)
@@ -413,14 +419,16 @@ rule scenario() -> Scenario
       t:tags()
       _
       pa:position!()
-      k:keyword((env.keywords().scenario_outline)) ":" _ n:not_nl() _ nl_eof()
+      k:keyword((env.keywords().scenario_outline)) ":" _ n:not_nl()? _ nl_eof()
+      d:description((&env.keywords().excluded_scenario_outline()))?
       s:steps()?
       e:examples()*
       pb:position!()
     {
         Scenario::builder()
             .keyword(k.into())
-            .name(n.to_string())
+            .name(n.unwrap_or_default().trim_end().to_string())
+            .description(d.flatten())
             .tags(t)
             .steps(s.unwrap_or_default())
             .examples(e)
@@ -473,7 +481,7 @@ pub(crate) rule tag_in_expr() -> String
     }
 
 pub(crate) rule tags() -> Vec<String>
-    = t:(tag() ** _) _ nl()* { t }
+    = t:(tag() ** __) _ nl()* { t }
     / { vec![] }
 
 rule rule_() -> Rule
@@ -481,7 +489,8 @@ rule rule_() -> Rule
       t:tags()
       _
       pa:position!()
-      k:keyword((env.keywords().rule)) ":" _ n:not_nl() _ nl_eof()
+      k:keyword((env.keywords().rule)) ":" _ n:not_nl()? _ nl_eof()
+      d:description((&env.keywords().excluded_rule()))?
       b:background()? nl()*
       s:scenarios()? nl()*
     //   e:examples()?
@@ -489,7 +498,8 @@ rule rule_() -> Rule
     {
         Rule::builder()
             .keyword(k.into())
-            .name(n.to_string())
+            .name(n.unwrap_or_default().trim_end().to_string())
+            .description(d.flatten())
             .tags(t)
             .background(b)
             .scenarios(s.unwrap_or_default())
@@ -507,10 +517,10 @@ pub(crate) rule scenarios() -> Vec<Scenario>
 pub(crate) rule feature() -> Feature
     = _ language_directive()?
       nl()*
-      t:tags() nl()*
+      t:tags()
       pa:position!()
-      k:keyword((env.keywords().feature)) ":" _ n:not_nl() _ nl()+
-      d:description()? nl()*
+      k:keyword((env.keywords().feature)) ":" _ n:not_nl()? _ nl_eof()
+      d:description((&env.keywords().excluded_feature()))?
       b:background()? nl()*
       s:scenarios() nl()*
       r:rules() pb:position!()
@@ -522,7 +532,7 @@ pub(crate) rule feature() -> Feature
             Ok(Feature::builder()
                 .keyword(k.into())
                 .tags(t)
-                .name(n.to_string())
+                .name(n.unwrap_or_default().trim_end().to_string())
                 .description(d.flatten())
                 .background(b)
                 .scenarios(s)
@@ -546,6 +556,8 @@ pub(crate) rule tag_operation() -> TagOperation = precedence!{
 
 #[cfg(test)]
 mod test {
+    use std::{collections::HashMap, fs};
+
     use super::*;
 
     const FOO: &str = "# language: formal\r\n
@@ -621,7 +633,7 @@ Scenario: Meow
     fn smoke2() {
         let env = GherkinEnv::default();
         let d = env!("CARGO_MANIFEST_DIR");
-        let s = std::fs::read_to_string(format!("{}/tests/test.feature", d)).unwrap();
+        let s = fs::read_to_string(format!("{}/tests/test.feature", d)).unwrap();
         assert!(gherkin_parser::feature(&s, &env).is_ok());
     }
 
@@ -724,12 +736,19 @@ Rule: rule
         assert_eq!(
             feature.rules[0].scenarios[0].examples[0]
                 .table
+                .as_ref()
+                .unwrap()
                 .position
                 .line,
             26,
         );
         assert_eq!(
-            feature.rules[0].scenarios[0].examples[0].table.rows.len(),
+            feature.rules[0].scenarios[0].examples[0]
+                .table
+                .as_ref()
+                .unwrap()
+                .rows
+                .len(),
             3,
         );
         assert_eq!(feature.rules[1].position.line, 32);
@@ -746,5 +765,200 @@ Rule: rule
         println!("{:#?}", feature);
         assert_eq!(feature.scenarios.len(), 0);
         assert!(feature.description.is_none());
+    }
+
+    #[test]
+    fn fixture_good() {
+        // We cannot handle missing features very well yet.
+        let skip = ["empty.feature", "incomplete_feature_3.feature"];
+        let mut failed = 0;
+
+        let d = env!("CARGO_MANIFEST_DIR");
+        let files = fs::read_dir(format!("{}/tests/fixtures/data/good/", d)).unwrap();
+        for file in files {
+            let file = file.unwrap();
+            let filename = file.file_name();
+            let filename = filename.to_str().unwrap();
+            if filename.ends_with(".feature") {
+                if skip.contains(&filename) {
+                    continue;
+                }
+                let res = std::panic::catch_unwind(|| {
+                    let env = GherkinEnv::default();
+                    let input =
+                        fs::read_to_string(format!("{}/tests/fixtures/data/good/{}", d, filename,))
+                            .unwrap();
+                    let feature = gherkin_parser::feature(&input, &env).unwrap();
+                    let fixture = fs::read_to_string(format!(
+                        "{}/tests/fixtures/data/good/{}.ast.ndjson",
+                        d, filename,
+                    ))
+                    .unwrap();
+
+                    check_ast(&feature, &fixture);
+                });
+                if res.is_err() {
+                    failed += 1;
+                    println!("{}", filename);
+                }
+            }
+        }
+
+        assert_eq!(failed, 0, "{} fixtures have failed", failed);
+    }
+
+    #[test]
+    fn fixture_fail() {
+        let d = env!("CARGO_MANIFEST_DIR");
+        let files = fs::read_dir(format!("{}/tests/fixtures/data/bad/", d)).unwrap();
+        for file in files {
+            let file = file.unwrap();
+            let filename = file.file_name();
+            let filename = filename.to_str().unwrap();
+            if filename.ends_with(".feature") {
+                let res = std::panic::catch_unwind(|| {
+                    let env = GherkinEnv::default();
+                    let input =
+                        fs::read_to_string(format!("{}/tests/fixtures/data/bad/{}", d, filename,))
+                            .unwrap();
+                    gherkin_parser::feature(&input, &env).unwrap()
+                });
+
+                assert!(res.is_err(), "{}: {:?}", filename, res.unwrap());
+            }
+        }
+    }
+
+    // TODO: Actually generate `.ndjson` file from the AST to fully assert
+    //       fixtures.
+    fn check_ast(parsed: &Feature, ast_parsed: &str) {
+        let d: HashMap<String, serde_json::Value> = serde_json::from_str(ast_parsed).unwrap();
+
+        let document = d
+            .get("gherkinDocument")
+            .expect("There is no document in the file");
+        let feature = document
+            .get("feature")
+            .expect("There is no feature in the document");
+        let children = feature.get("children");
+
+        if children.is_none() {
+            assert!(parsed.background.is_none());
+            assert_eq!(parsed.scenarios.len(), 0);
+            assert_eq!(parsed.rules.len(), 0);
+            return;
+        }
+
+        let mut backgrounds = 0;
+        let mut scenarios = 0;
+        let mut rules = 0;
+
+        let children = children.unwrap().as_array().unwrap();
+        for child in children {
+            if let Some(background) = child.get("background") {
+                let parsed_background = parsed.background.as_ref().unwrap();
+                let name = background.get("name").unwrap().as_str().unwrap();
+
+                assert_eq!(&parsed_background.name, name);
+
+                let steps = background.get("steps");
+
+                if steps.is_none() {
+                    assert_eq!(parsed_background.steps.len(), 0);
+                    continue;
+                }
+
+                let steps = steps
+                    .expect("steps")
+                    .as_array()
+                    .expect("Steps must be an array");
+                assert!(
+                    check_steps(&parsed_background.steps, steps),
+                    "Background steps are different from fixture",
+                );
+
+                backgrounds += 1;
+            } else if let Some(json_scenario) = child.get("scenario") {
+                assert!(
+                    check_scenario(&parsed.scenarios, json_scenario),
+                    "Scenario steps are different from fixture",
+                );
+
+                scenarios += 1;
+            } else if let Some(json_rule) = child.get("rule") {
+                let json_rule_scenarios = json_rule
+                    .get("children")
+                    .and_then(serde_json::Value::as_array)
+                    .map(|children| {
+                        children
+                            .iter()
+                            .filter_map(|child| child.get("scenario"))
+                            .collect::<Vec<_>>()
+                    });
+                let json_rule_name = json_rule.get("name").unwrap().as_str().unwrap();
+
+                let exists = parsed.rules.iter().any(|rule| {
+                    if rule.name == json_rule_name {
+                        return if let Some(json_scenarios) = &json_rule_scenarios {
+                            json_scenarios.len() == rule.scenarios.len()
+                                && json_scenarios.iter().all(|json_scenario| {
+                                    check_scenario(&rule.scenarios, json_scenario)
+                                })
+                        } else {
+                            rule.scenarios.is_empty()
+                        };
+                    }
+                    false
+                });
+
+                assert!(exists, "Scenario rules are different from fixture");
+
+                rules += 1;
+            } else {
+                panic!("Unknown child type: {:#?}", child);
+            }
+        }
+
+        if parsed.background.is_some() {
+            assert_eq!(1, backgrounds);
+        } else {
+            assert_eq!(0, backgrounds);
+        }
+
+        assert_eq!(parsed.scenarios.len(), scenarios);
+        assert_eq!(parsed.rules.len(), rules);
+    }
+
+    fn check_scenario(parsed: &[Scenario], json: &serde_json::Value) -> bool {
+        let json_steps = json.get("steps").and_then(serde_json::Value::as_array);
+        let json_scenario_name = json.get("name").unwrap().as_str().unwrap();
+
+        parsed.iter().any(|parsed| {
+            if parsed.name == json_scenario_name {
+                return if let Some(json_steps) = json_steps {
+                    check_steps(&parsed.steps, json_steps)
+                } else {
+                    parsed.steps.is_empty()
+                };
+            }
+            false
+        })
+    }
+
+    fn check_steps(parsed: &[Step], json: &[serde_json::Value]) -> bool {
+        if parsed.len() != json.len() {
+            return false;
+        }
+
+        for (step, json) in parsed.iter().zip(json) {
+            if step.keyword != json.get("keyword").unwrap().as_str().unwrap() {
+                return false;
+            }
+            if step.value != json.get("text").unwrap().as_str().unwrap() {
+                return false;
+            }
+        }
+
+        true
     }
 }